+++ apps/user/src/components/signup/Terms/Terms.tsx
@@ -3,15 +3,50 @@ import { Dispatch, SetStateAction, useEffect, useState } from 'react';
 import { color, font } from '@maru/theme';
 import { flex } from '@maru/utils';
 import styled from 'styled-components';
-import useTerms from './Terms.hook';
 
 interface PropsType {
     setTermsAgree: Dispatch<SetStateAction<boolean>>;
 }
 
 const Terms = ({ setTermsAgree }: PropsType) => {
-    const { allCheck, checkOne, checkTwo, allCheckEvent, checkOneEvent, checkTwoEvent } =
-        useTerms(setTermsAgree);
+    const [allCheck, setAllCheck] = useState(false);
+    const [checkOne, setChecktOne] = useState(false);
+    const [checkTwo, setChecktTwo] = useState(false);
+
+    const allCheckEvent = () => {
+        if (allCheck === false) {
+            setChecktOne(true);
+            setChecktTwo(true);
+        } else {
+            setChecktOne(false);
+            setChecktTwo(false);
+        }
+    };
+
+    const checkOneEvent = () => {
+        if (checkOne === false) {
+            setChecktOne(true);
+        } else {
+            setChecktOne(false);
+        }
+    };
+
+    const checkTwoEvent = () => {
+        if (checkTwo === false) {
+            setChecktTwo(true);
+        } else {
+            setChecktTwo(false);
+        }
+    };
+
+    useEffect(() => {
+        if (checkOne === true && checkTwo === true) {
+            setAllCheck(true);
+            setTermsAgree(true);
+        } else {
+            setAllCheck(false);
+        }
+    }, [checkOne, checkTwo]);
 
     return (
         <StyledTerms>
