+++ packages/maru-hooks/src/useInput.ts
@@ -1,4 +1,4 @@
-import { useState, useCallback, ChangeEvent, useMemo } from 'react';
+import { useState, useCallback, useMemo, ChangeEventHandler } from 'react';
 import debounce from 'lodash/debounce';
 
 interface PropsType {
@@ -29,14 +29,15 @@ const useInput = ({ initialValue = '', useDebounce = false, debounceTimeout = 15
         [handleSetDebounceValue, useDebounce],
     );
 
-    const onChange = useCallback(
-        (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-            handleSetValue(e.target.value);
-        },
-        [handleSetValue],
-    );
+    const handleInputDataChange: ChangeEventHandler<HTMLInputElement | HTMLTextAreaElement> =
+        useCallback(
+            (e) => {
+                handleSetValue(e.target.value);
+            },
+            [handleSetValue],
+        );
 
-    return { value, setValue, onChange, debouncedValue };
+    return { value, setValue, handleInputDataChange, debouncedValue };
 };
 
 export default useInput;
